Make mips241 compilable on Windows
In Windows, the symbols defined for a shared library needs to be
explicitly exported/imported using the directive `__declspec(dll*)`.
After this, the .dll is created containing the data and a .lib file
containing the symbol names is created for linking. I added those so
the program can be built and the frontend can load the right library.

I defined the macro that expands to the right directive on Windows
compilers (note that I'm not sure how ICC Windows works), and expands to
nothing on Linux. So it can still normally run on Linux platforms.

I could use CMake to handle this: by adding the following to
src/CMakeLists.txt:
```
include(GenerateExportHeader)
generate_export_header(mips241
    BASE_NAME mips241
    EXPORT_MACRO_NAME mips241_EXPORT
    EXPORT_FILE_NAME ${CMAKE_CURRENT_SOURCE_DIR}/common/mips241_export.h
    STATIC_DEFINE mips241_BUILT_AS_STATIC
)
```
and add these lines in src/common/defs.h:
```
define mips241_EXPORTS
include "mips241_export.h"
```

However this only works for C++ codes for older versions of CMake.